api/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.1
django-jazzmin==3.0.1
django-redis==5.4.0
django-stubs==5.1.0
django-stubs-ext==5.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
packaging==24.1
Pillow==11.0.0
prompt_toolkit==3.0.47
requests==2.31.0
djangorestframework-simplejwt==5.3.0
numpy==1.26.4
psycopg2==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.20.0
sentry-sdk==2.17.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
channels==4.0.0
channels-redis==4.1.0
